cf-geminiimagegenerator/requirements.txt
functions-framework==3.*
flask>=2.0.0
google-genai>=0.7.0
google-cloud-storage
python-dotenv>=0.21.0
requests>=2.31.0